requirements.txt
python-telegram-bot==21.3
pytube==15.0.0
tqdm==4.62.3
pyrogram==2.0.10
TgCrypto==1.2.4
ntplib==0.4.0
python-dotenv==1.0.1
requests==2.32.3